Convert to multi-module project with Gradle build on Java 8.
The project is building successfully with Gradle now.
I also added the source code for the PluginLoaderAPI
which was sent to me by way of email from the original
auther (Michel Buffa) <micbuffa@gmail.com>. He sent
it to me so that I could modify the code in
"fr.unice.plugin.Plugin" which was preventing the
application from running on JVM versions higher than
version 1.6. The problem here was that Java 7 added
a method to java.awt.Window called getType() which
conflicted with fr.unice.plugin.Plugin.getType(). The
solution was rather simple, just rename that method
to getPluginType(). Making this change allows
the application to compile and run on Java 8.

The application needs to be reviewed and redesigned to
better share code between the various modules. Ideally
it would be good to get this application off of the
current plugin infrastruction all together and replace
it using OSGi.